transition

+++ tgs_alone/Scene/EndScene.h
@@ -12,7 +12,7 @@ private:
 	int tran_img;         // トランジション画像
 	int star_cnt;  // 星の回転用
 	bool tran_flg;        // 画面遷移 false:しない true:する
-	float transition;     // トランジションの値保存用
+	int transition;     // トランジションの値保存用
 	
 public:
 	static int cnt;  // 終了までのカウント用変数
@@ -43,6 +43,6 @@ public:
 	// 星の回転処理
 	void StarAnim();
 
-	// トランジション処理
+	// 画面切り替えアニメーション
 	void Transition();
 };
