+++ tgs_alone/Scene/ResultScene.h
@@ -22,7 +22,7 @@ private:
 	int combo[5];      // コンボデータ
 	int star_cnt;      // 星の回転用
 	bool tran_flg;        // 画面遷移 false:しない true:する
-	float transition;     // トランジションの値保存用
+	int transition;     // トランジションの値保存用
 
 public:
 
@@ -52,6 +52,6 @@ public:
 	// 星の回転処理
 	void StarAnim();
 
-	// トランジション処理
+	// 画面切り替えアニメーション
 	void Transition();
 };
